Adjusting structure to be bundler friendly
Prior to this commit if I ran `bundle exec rake` I got the following
error:

```console
bundler: failed to load command: rake
Gem::Exception: can't find executable rake for gem rake. rake is not
  currently included in the bundle, perhaps you meant to add it to your
  Gemfile?
```

Once I added rake to the gemspec, I began encountering additional errors
related to path and a constant being declared as both a `module` and
a `class`.

This is the minimum to tidy that up.